更新 , 去掉 warning , 由于 NSTimer 会被 触摸打包后响应的 UIEvent 阻塞 , 所以用 GCD 定时器替换掉 NSTimer .

+++ Rainville/Rainville/Classes/Modules/Controllers + Views/Main/CCMainViewController.m
@@ -102,7 +102,7 @@
 
 #pragma mark - CCCellTimerDelegate
 - (void)ccCellTimerWithSeconds:(NSInteger)integerSeconds {
-    CCLog(@"_CC_MAIN_TIMER_%ld",integerSeconds);
+    CCLog(@"_CC_MAIN_TIMER_%ld",(long)integerSeconds);
     ccWeakSelf;
     [_handler ccSetAutoStopWithSeconds:integerSeconds withBlock:^(BOOL isSucceed, id item) {
         if ([item isKindOfClass:[NSString class]]) {
@@ -151,7 +151,7 @@
 
 - (void) ccClickedAction : (NSInteger) integerIndex
                  withKey : (NSString *) stringKey {
-    CCLog(@"_CC_CLICKED_%ld_KEY_%@_",integerIndex,stringKey);
+    CCLog(@"_CC_CLICKED_%ld_KEY_%@_",(long)integerIndex,stringKey);
     ccWeakSelf;
     [_handler ccSetAudioPlayerWithVolumeArray:_dictionaryTheme[stringKey] withCompleteHandler:^{
         CCLog(@"_CC_PLAY_SUCCEED_");
